feat: improve auth account and order handling

+++ src/main/java/org/example/shopping/model/CartItemResponse.java
@@ -1,4 +1,5 @@
 package org.example.shopping.model;
+import java.math.BigDecimal;
 
 /** Dòng sản phẩm hiển thị trong giỏ hàng. */
 public class CartItemResponse {
@@ -9,11 +10,11 @@ public class CartItemResponse {
     /** Tên sản phẩm hiển thị trong giỏ. */
     private String name;
     /** Đơn giá hiện tại lấy từ database. */
-    private Double price;
+    private BigDecimal price;
     /** Số lượng khách đã chọn. */
     private Integer quantity;
     /** Thành tiền của dòng giỏ: price × quantity. */
-    private Double subtotal;
+    private BigDecimal subtotal;
 
     public Integer getProductId() { return productId; }
     public void setProductId(Integer productId) { this.productId = productId; }
@@ -21,10 +22,10 @@ public class CartItemResponse {
     public void setCode(String code) { this.code = code; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
     public Integer getQuantity() { return quantity; }
     public void setQuantity(Integer quantity) { this.quantity = quantity; }
-    public Double getSubtotal() { return subtotal; }
-    public void setSubtotal(Double subtotal) { this.subtotal = subtotal; }
+    public BigDecimal getSubtotal() { return subtotal; }
+    public void setSubtotal(BigDecimal subtotal) { this.subtotal = subtotal; }
 }
